Complete react exercise
The weird-but-pleasant thing about this is the way I suddenly don't
need to worry about the lifetimes of the callbacks anymore--I didn't
think that I'd actually solved that last time, but apparently I did.

These changes are all about matching the expected semantics--I'd
managed to get a few things wrong.