Partial work computing the pythagorean triplet at compile time
Unfortunately, we're blocked here for now: const functions cannot
yet contain if statements, and we need that to perform the recursive
search for a valid triplet. It's a race, now: will rust's const fn
support evolve faster than the pythagorean triplet test suite, or
not? There are open issues for both.

I'll likely just come back and try this again in a few months; maybe
we'll have a resolution then.